requirements.txt
#python packages to install
python-dotenv==1.0.1
pymongo==4.6.3